Upgrading SV health (nav) support
