+++ src/navigation/ephemeris/health/bds.rs
@@ -1,0 +1,6 @@
+use bitflags::bitflags;
+
+use num_derive::FromPrimitive;
+
+#[cfg(feature = "serde")]
+use serde::Serialize;
