+++ src/navigation/ephemeris/health/mod.rs
@@ -1,0 +1,6 @@
+pub mod bds;
+pub mod gal;
+pub mod geo;
+pub mod glonass;
+pub mod gps;
+pub mod irnss;
